% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{adj_se.default}
\alias{adj_se.default}
\title{Prints the adjusted standard erorr model.}
\usage{
\method{adj_se}{default}(mod, transformation)
}
\arguments{
\item{mod}{A regression model.}

\item{transformation}{Either a function to be transform a vector of
standard errors or a vector of new standard errors to replace the model
standard errors}
}
\value{
A custom S3 representation of the regression model
  with adjusted standard errors.
}
\description{
Prints the adjusted standard erorr model.
}
