% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{adj_se}
\alias{adj_se}
\title{Simple adjustment of model standard errors}
\usage{
adj_se(mod, f)
}
\arguments{
\item{mod}{A (linear) model.}

\item{f}{A function that transforms the model's standard errors.}
}
\value{
a modified version of the inputted model
}
\description{
Simple adjustment of model standard errors
}
