% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Generates Latex Appendix}
\usage{
append(models, dir)
}
\arguments{
\item{models}{A list of models generated by \code{build} or \code{describe}.}
}
\description{
Creates a LaTeX appendix of regression summaries and descriptive statistics.
  A directory containing all the relevant files
}
\examples{

data('freeny')
mod <- lm(y ~ lag.quarterly.revenue + price.index + income.level +
            market.potential, data = freeny)
mod2 <- lm(y ~ lag.quarterly.revenue + price.index + income.level,
          data = freeny)

tex1 <- build(mod, label='f1', silent=TRUE)
tex2 <- build(mod2, label='f2', silent=TRUE)
d <- paste(getwd(), 'rchitEx', sep='/')
append(list(tex1, tex2), d)
## <-- Latex files outputted to dir --> ##

}
