% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build.function}
\alias{build.function}
\title{builder function}
\usage{
\method{build}{function}(..., dep_names = NA, indep_names = NA,
  note = "", title = "Model results", max_precision = 3, path = NA,
  rse = FALSE, silent = FALSE, landscape = FALSE, report = "p",
  stats = "oraf", pre_stats = NA, md = NA, header = TRUE,
  label = "table", sig = NA)
}
\description{
builder function
}
