% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Formats Regression Tables to Text, LaTeX, and HTML}
\usage{
build(..., dep_names = NULL, indep_names = NULL, note = "",
  title = "Model results", max_precision = 3, path = NULL,
  rse = FALSE, silent = FALSE, landscape = FALSE, report = "p",
  stats = "oraf", pre_stats = NULL, md = NULL, header = TRUE,
  label = "table", sig = NULL, as_table = TRUE,
  grouped_label = NULL)
}
\arguments{
\item{...}{Regression models to be included in Table.  Table will display models in the same order as provided.}

\item{dep_names}{Character vector of dependent variable names. Vector should be listed in the same order
that the models will be displayed (from left to right). If NULL, the columns will be numbers (e.g. (1), (2)....)}

\item{indep_names}{List of names associating independent variable names with desired names in table.}

\item{note}{Optional note displayed in bottom row of the table.}

\item{title}{Title of the table.}

\item{max_precision}{Maximum number of digits in a table cell.}

\item{path}{Output path for the tex file. NA for no tex output.}

\item{rse}{Transforms to robust standard errors if true. Uses HC1 like STATA.}

\item{silent}{No text output if true.}

\item{landscape}{If true, the Latex table will be landscaped.}

\item{report}{Test statistic to be provided along with the coefficient.
Must be either "p", "t", or "ste". Default value is "p".}

\item{stats}{Fit characteristics to be provided at bottom of table. Must be a string of reporters.
Options include "o" for number of observations, "r" for R2, "a" for adjusted R2 and "f" for f-statistic.
The inputted string determines the order of outputted reporters. See vignette for full list of
possible values. Ex \code{stats='oraf'}.}

\item{pre_stats}{List of optional reporters to be provided above normal fit characteristics.
The list must be formatted as the name of the reporter followed by a a vector of values.
Ex \code{pre_stats=list('R.St.E' = c('True', 'True'))}.}

\item{md}{Allows for outputting in either latex ("latex") or html ("html") for markdown formatting.
The Markdown chunk must be set to \code{results = "asis"}.}

\item{header}{Includes RCHITEX header as a Latex comment if true.}

\item{label}{Latex label.}

\item{sig}{List containing the associations between significance symbols and cut-off p values.
List must be in increasing order of cutoff values (ex \code{list('***' = 0.01, '**' = 0.05, '*' = 0.1)}).}

\item{as_table}{True values wrap the underlying Latex Tabular object in a table.}

\item{grouped_label}{Optional label printed above model names that group models together.
Expected format is a list of the name and a vector of the start and end columns.}
}
\value{
Invisible return containing rchitex object.
}
\description{
Given a linear regression model, \code{build} constructs a formatted
  text table for standard output as well as a Latex table that is
  outputted to an provided file path.
}
\examples{
library(rchitex)
data(swiss)

mod1 <- lm(data=swiss, Fertility ~ Agriculture + Education)
mod2 <- lm(data=swiss, Fertility ~ Agriculture + Education + Infant.Mortality + Catholic + Examination)
lmod <- glm(data=swiss, I(Fertility > mean(Fertility)) ~ Agriculture +
              Education + Infant.Mortality + Catholic + Examination,
            family=binomial(link='logit'))

indep_names <- list('Agriculture' = "Agriculture share", 'Education' = 'Total education',
                    'Infant.Mortality' = 'Infant Mortality',
                    'Catholic' = 'Catholic share', 'Examination' = 'Exam')
dep_names <- c('Fert.', 'Fert.', 'Fert.')
grouped_label <- list('OLS' = c(1,2), 'logit' = 3)
pre_stats <- list('Full dataset' = c('No', 'Yes', 'Yes'))
sig <- list('***' = 0.001, '**' = 0.025, "'"=0.15)
}
