% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textify.R
\name{center}
\alias{center}
\title{Centers a phrase within a n-length string. Eg '"Ben", 7' -> "  Ben  "}
\usage{
center(word, total_spaces, sepr = "")
}
\description{
Centers a phrase within a n-length string. Eg '"Ben", 7' -> "  Ben  "
}
