% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textify.R
\name{column_widths}
\alias{column_widths}
\title{Calculates the maximum number of spaces needed in each column}
\usage{
column_widths(idvn_names, fits, dv_names, coefs, sigs)
}
\arguments{
\item{idvn_names}{vector of independent variable names}

\item{dv_names}{vector or dependent variable names}

\item{coefs}{vector of coefficients}
}
\description{
Calculates the maximum number of spaces needed in each column
}
