% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tex.R
\name{gg_tex}
\alias{gg_tex}
\title{Wraps ggplot as Latex Objects}
\usage{
gg_tex(ggOb, label, caption = "Figure")
}
\arguments{
\item{ggOb}{A ggplot.}

\item{label}{Intended Latex label for plot figure.}

\item{caption}{Intended Latex title for plot figure.}
}
\description{
gg_tex allows ggplots to be included in automated Latex appendixes constructing through \code{\link{appendize}}.
}
\examples{
df <- data.frame("first" = c(4,5,6), "second" = c(7,5,3))
g <- ggplot(data=df) + geom_point(aes(x=first, y=second))
rchitex::appendize(gg_tex(g, label='ggFig', caption='My ggplot'), path=getwd())

}
