% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textify.R
\name{grouper}
\alias{grouper}
\title{created centered multi-column grouped labels with approporiate underlining}
\usage{
grouper(gl_vector, col_widths)
}
\description{
created centered multi-column grouped labels with approporiate underlining
}
