% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textify.R
\name{model2text}
\alias{model2text}
\title{Convert a regression-like model to a text table format}
\usage{
model2text(
  coefs,
  reporter,
  fits,
  sigs,
  idvn,
  max_precision,
  sig_levels,
  title = "Regression Results",
  note = "",
  dn,
  grouped_label = NULL
)
}
\arguments{
\item{coefs}{vector of coefficients}

\item{idvn}{vector of independent variable names}

\item{max_precision}{int > 0, number of digits to display (e.g. 0.XXXX = 4)}

\item{sig_levels}{list of significance levels and the corresponding symbols}

\item{note}{(UNSUPPORTED) string note at the bottom of the table}
}
\description{
Convert a regression-like model to a text table format
}
