% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{nobs.rse}
\alias{nobs.rse}
\title{Extract number of observations}
\usage{
\method{nobs}{rse}(object, ...)
}
\arguments{
\item{object}{RSE model.}

\item{...}{Additional parameters}
}
\description{
Extract number of observations
}
