% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textify.R
\name{p_value_row}
\alias{p_value_row}
\title{Produces the row of p values and cut offs. Located at bottom of table}
\usage{
p_value_row(sig_levels)
}
\description{
Produces the row of p values and cut offs. Located at bottom of table
}
