% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{print.adj_se}
\alias{print.adj_se}
\title{Prints adjusted SE model#'}
\usage{
\method{print}{adj_se}(x, ...)
}
\arguments{
\item{x}{Adjusted SE model.}

\item{...}{Additional parameters}
}
\description{
Prints adjusted SE model#'
}
