% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{print.rchReg}
\alias{print.rchReg}
\title{print statement for build}
\usage{
\method{print}{rchReg}(x, ...)
}
\arguments{
\item{x}{Rchitex model to be printed.}

\item{...}{Additional parameters}
}
\description{
print statement for build
}
