% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{print.rcTable}
\alias{print.rcTable}
\title{print statement for build}
\usage{
\method{print}{rcTable}(x, ...)
}
\arguments{
\item{x}{Rchitex summary table to be printed.}

\item{...}{Additional parameters}
}
\description{
print statement for build
}
