% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchitex.R
\docType{package}
\name{rchitex}
\alias{rchitex}
\alias{rchitex-package}
\title{rchitex: Streamlining the transition from R to LaTeX}
\description{
The rchitex package is designed to facilitate the creation
of LaTeX articles from R models. rchitex provides
three main functions, \code{\link{build}} for regression
modls, \code{\link{describe}} for summary statistics,
and \code{\link{appendize}} and \code{\link{gg_tex}} for creating an entire LaTeX
appendix.
}
\details{
rchitex's tex outputting allows for elegant console data investigation.
Data can be summarized and outputted with a wide range of customizations.

Along with text output, rchitex supports latex and html output, making the
package optimal for use in Markdown files.
}
