% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{rse.default}
\alias{rse.default}
\title{Simple adjustment of model standard errors.}
\usage{
\method{rse}{default}(m)
}
\arguments{
\item{m}{A linear model.}
}
\value{
A custom S3 representation of the linear model with robust standard errors.
}
\description{
Allows model standard errors to be adjusted
 prior to table output
}
