% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{rse.function}
\alias{rse.function}
\title{builder function}
\usage{
\method{rse}{`function`}(...)
}
\arguments{
\item{...}{Shouldn't be used so arguments not specified}
}
\description{
builder function
}
