% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{rse}
\alias{rse}
\title{Simple adjustment of model standard errors}
\usage{
rse(mod)
}
\description{
Allows model standard errors to be adjusted
 prior to table output
 @param mod A (linear) model
 @return a modified version of the inputted model
}
