% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{summary.adj_se}
\alias{summary.adj_se}
\title{Summarizes adjusted SE model}
\usage{
\method{summary}{adj_se}(object, ...)
}
\arguments{
\item{object}{Adjusted SE model.}

\item{...}{Additional parameters}
}
\description{
Summarizes adjusted SE model
}
