% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{summary.rse}
\alias{summary.rse}
\title{Summarizes RSE models}
\usage{
\method{summary}{rse}(object, ...)
}
\arguments{
\item{object}{RSE model.}

\item{...}{Additional parameters}
}
\description{
Summarizes RSE models
}
