% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textify.R
\name{write_idv}
\alias{write_idv}
\title{Writes 2 lines of reg table for a given dependent variable
note - function is vectorized}
\usage{
write_idv(iv, idvn, coefs, sigs, col_widths, reporter)
}
\description{
Writes 2 lines of reg table for a given dependent variable
note - function is vectorized
}
