% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build.function}
\alias{build.function}
\title{builder function}
\usage{
\method{build}{function}(...)
}
\arguments{
\item{...}{Shouldn't be used so arguments not specified}
}
\description{
builder function
}
