% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Formats Tables to Text, LaTeX, and HTML}
\usage{
build(..., dep_names = NA, indep_names = NA, note = "",
  title = "Model results", max_precision = 3, path = NA,
  rse = FALSE, silent = FALSE, landscape = FALSE, report = "p",
  stats = "oraf", pre_stats = NA, md = NA, header = TRUE,
  label = "table")
}
\arguments{
\item{...}{Regression models to be included in Table.  Table will display models in the same order as provided.}

\item{dep_names}{List associating dependant variable names with desired names in table.}

\item{indep_names}{List of names associating independant variable names with desired names in table.}

\item{note}{Optional note displayed in bottom row of the table.}

\item{title}{Title of the table.}

\item{max_precision}{Maximum number of digits in a table cell.}

\item{path}{Output path for the tex file. NA for no tex output.}

\item{rse}{Transforms to robust standard errors if true. Uses HC1 like STATA.}

\item{silent}{No text output if true.}

\item{landscape}{Formats tex table as landscape if true. If landscape is true, make sure to add to include
\code{\documentclass{lscape}} in your LaTeX code.}

\item{report}{Test statistic to be provided along with the coefficient.
Must be either "p", "t", or "ste". Default value is "p".}

\item{stats}{Fit characteristics to be provided at bottom of table. Must be a string of reporters.
Options include "o" for number of observations, "r" for R2, "a" for adjusted R2 and "f" for f-statistic.
The inputted string determines the order of outputted reporters. Ex \code{stats='oraf'}.}

\item{pre_stats}{List of optional reporters to be provided above normal fit characteristics.
The list must be formatted as the name of the reporter followed by a a vector of values.
Ex \code{pre_stats=list('R.St.E' = c('True', 'True'))}.}

\item{md}{The markdown parameter allows for outputting in either latex ("latex") or html ("html") for markdown formatting.
The Markdown chunk must be set to \code{results = "asis"}.}

\item{header}{Outputs RCHITEX header if true.}

\item{label}{Latex label.}
}
\value{
Invisible return containing a list of model attributes.
}
\description{
Given a linear regression model, \code{build} constructs a formatted
  text table for standard output as well as a Latex table that is
  outputted to an provided file path.
}
\examples{
data('freeny')
mod <- lm(y ~ lag.quarterly.revenue + price.index + income.level +
           market.potential, data = freeny)
mod2 <- lm(y ~ lag.quarterly.revenue + price.index + income.level,
          data = freeny)
d <- list('price.index' = 'Price index', 'income.level' = 'Income level',
         'market.potential' = 'Market potential',
         'lag.quarterly.revenue' = 'Lagged rev')
pre <- list('R. St. E' = c('True', 'True'))
build(mod, mod2, rse = T, indep_names = d, pre_stats = pre)

}
