% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{nobs.adj_se}
\alias{nobs.adj_se}
\title{Extracts number of observations of adjusted SE model}
\usage{
\method{nobs}{adj_se}(object, ...)
}
\arguments{
\item{object}{Adjusted SE model.}

\item{...}{Additional parameteres}
}
\description{
Extracts number of observations of adjusted SE model
}
