% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{print.rcTable}
\alias{print.rcTable}
\title{print statement for build}
\usage{
\method{print}{rcTable}(m)
}
\arguments{
\item{m}{Rchitex summary table to be printed.}
}
\description{
print statement for build
}
