% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_adjuster.R
\name{summary.adj_se}
\alias{summary.adj_se}
\title{Summarizes adjusted SE model}
\usage{
\method{summary}{adj_se}(m)
}
\arguments{
\item{m}{Adjusted SE model.}
}
\description{
Summarizes adjusted SE model
}
